requirements.txt
pyTelegramBotAPI==4.31.0
python-dotenv==1.2.1